LLM part/requirements.txt
torch
torchvision
streamlit
openai
pillow
